Split zip_iterator.pass into 8 tests
This is done to avoid having an excessive number of kernels in a single test which
exhausts all of the available device memory in some discrete client GPUs.
